refactor: Extract controller funcs and globals from main
