+++ project/lesotho/lesotho/global/glo.go
@@ -1,0 +1,9 @@
+package global
+
+import (
+	"lesotho/acl"
+	ns "lesotho/namespace"
+)
+
+var Acl *acl.ACL
+var Nss *ns.NamespaceStore
